debug

+++ Train.py
@@ -57,7 +57,7 @@ class SimpleLossCompute:
         x = self.generator(x)
         print(x.contiguous().view(-1, x.size(-1)).type())
         print(y.contiguous().view(-1).float().type())
-        loss = self.criterion(x.contiguous().view(-1, x.size(-1)), y.contiguous().view(-1).float()) / norm
+        loss = self.criterion(x.contiguous().view(-1, x.size(-1)), y.contiguous().view(-1).float()) / norm.float()
         loss.backward()
         if self.opt is not None:
             self.opt.step()
